Implement expr variables
To do so, check that the kind of the variable actually matches the kind
we're trying to parse and return None if it doesnt match. Currently the
lookup is done in the slowest way possible, but it shouldn't be a big
deal (this only happens while parsing molt files anyways, so its not a
performance issue).