Implement RequiresRule for CmpSyn
In the previous commit, I introduced a bug in which I forgot to compare
the rules for the visibility in the ImplItem <-> Item CmpSyn
implementation. To avoid this in the future, I added a generic to
CmpSyn. This generic is `DoesNotRequireRule` by default but will be
defined as `RequiresRule` when deriving the `CmpSyn` trait with the
`#[requires_rule]` attribute present. The trick is that the `cmp_syn`
method on `Matcher` only exists for `DoesNotRequireRule` impls of
`CmpSyn`, and calling a more specific method  (`cmp_syn_with_rule`)
is required to use the latter type of impl. This ensures checking the
rule cannot be forgotten.