Split where clause off from Generics.
In the existing logic, we had to modify the parsed generics after the
parsing, leading to uglier code than I'd like (because it means we need
to parse patterns and then match on them and treat the case of variables
specially.)

After the split, things should be 1. easier to parse and 2. easier to
split into separate node types (i.e. allow matching on generics or where
clauses separately).