Add proper Ctrl+C handling for temp modifications.
Before, pressing Ctrl+C while a `--check` run was in progress would just
leave the temporary modification (which is being made in order to run
the check script on the modified code base) lying around. These changes
make sure that any modifications made as temporary modifications are
tracked and restored if Ctrl+C is caught.